Fully functional Networking Class and Queque handler
# Networking
I created a networking Class to send (asynchronous) message to the
server in a simply way using Json.
You can use the networking.feedback var to store a function and resolve
it in the unity update function, in this mode you can search a
gameObject without get into thread safe exeption.

# GlobaQueque
I don't know if is better to create a queque for evry gameObject  or a
global queque like now. I not to think heavily on this thing.. i really
don't know.